+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   `fstatat64()`, `lstat64()`, `open64()`, `openat64()` and `stat64()` for
   [musl] world
 - Intercept the [FreeBSD]'s functions `chflags()`, `chflagsat` and `lchflags()`
+- The internal function `path_resolution()` expands all symlinks
 
 ### Changed